requirements.txt
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.2
ddt==1.2.0
defusedxml==0.5.0
Django==1.11
django-precise-bbcode==1.2.10
django-taggit==0.22.2
easy-thumbnails==2.5
idna==2.7
nose==1.3.7
oauthlib==2.1.0
Pillow==5.2.0
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23
